+++ prefect_hex/models/project.py
@@ -123,8 +123,8 @@ class ProjectStatusResponsePayload(BaseModel):
         alias="elapsedTime",
         description="Total elapsed time for the project run in milliseconds.",
     )
-    end_time: datetime = Field(
-        default=...,
+    end_time: Optional[datetime] = Field(
+        default=None,
         alias="endTime",
         description="UTC timestamp of when the project run finished.",
     )