+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `trigger_project_run_and_wait_for_completion` flow - [#15](https://github.com/PrefectHQ/prefect-hex/pull/15)
+
 ### Changed
 
 ### Deprecated
@@ -16,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+
+- Make `end_time` an optional field in `ProjectStatusResponsePayload` - [#14](https://github.com/PrefectHQ/prefect-hex/pull/14)
 
 ### Security